# specs/contract/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "SDD Starter",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/sys/dicts": {
      "get": {
        "tags": [
          "sys_dict"
        ],
        "summary": "List Dicts",
        "operationId": "listSysDicts",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 50,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DictOut"
                  },
                  "title": "Response Listsysdicts"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-trace": [
          "FR-6.5.1"
        ]
      },
      "post": {
        "tags": [
          "sys_dict"
        ],
        "summary": "Create Dict",
        "operationId": "createSysDict",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-trace": [
          "FR-6.5.1"
        ]
      }
    },
    "/api/sys/dicts/{dict_id}": {
      "put": {
        "tags": [
          "sys_dict"
        ],
        "summary": "Update Dict",
        "operationId": "updateSysDict",
        "parameters": [
          {
            "name": "dict_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dict Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-trace": [
          "FR-6.5.1"
        ]
      },
      "delete": {
        "tags": [
          "sys_dict"
        ],
        "summary": "Delete Dict",
        "operationId": "deleteSysDict",
        "parameters": [
          {
            "name": "dict_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Dict Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-trace": [
          "FR-6.5.1",
          "NFR-6.1"
        ]
      }
    },
    "/api/sys/dicts/{type_code}/items": {
      "get": {
        "tags": [
          "sys_dict"
        ],
        "summary": "List Items",
        "operationId": "listSysDictItems",
        "parameters": [
          {
            "name": "type_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Type Code"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "default": 0,
              "title": "Offset"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DictItemOut"
                  },
                  "title": "Response Listsysdictitems"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-trace": [
          "FR-6.5.2"
        ]
      }
    },
    "/api/sys/dicts/items": {
      "post": {
        "tags": [
          "sys_dict"
        ],
        "summary": "Create Item",
        "operationId": "createSysDictItem",
        "parameters": [
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictItemCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-trace": [
          "FR-6.5.2"
        ]
      }
    },
    "/api/sys/dicts/items/{item_id}": {
      "put": {
        "tags": [
          "sys_dict"
        ],
        "summary": "Update Item",
        "operationId": "updateSysDictItem",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DictItemUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DictItemOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-trace": [
          "FR-6.5.2"
        ]
      },
      "delete": {
        "tags": [
          "sys_dict"
        ],
        "summary": "Delete Item",
        "operationId": "deleteSysDictItem",
        "parameters": [
          {
            "name": "item_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Item Id"
            }
          },
          {
            "name": "authorization",
            "in": "header",
            "required": false,
            "schema": {
              "type": "string",
              "default": "",
              "title": "Authorization"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-trace": [
          "FR-6.5.2",
          "NFR-6.1"
        ]
      }
    },
    "/api/sys/dicts/public/{type_code}/items": {
      "get": {
        "tags": [
          "sys_dict"
        ],
        "summary": "Get Enabled Items",
        "operationId": "getEnabledSysDictItems",
        "parameters": [
          {
            "name": "type_code",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Type Code"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/DictItemOut"
                  },
                  "title": "Response Getenabledsysdictitems"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "x-trace": [
          "FR-6.5.3"
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "DictCreate": {
        "properties": {
          "code": {
            "type": "string",
            "minLength": 1,
            "title": "Code"
          },
          "name": {
            "type": "string",
            "minLength": 1,
            "title": "Name"
          },
          "remark": {
            "type": "string",
            "title": "Remark",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "code",
          "name"
        ],
        "title": "DictCreate"
      },
      "DictItemCreate": {
        "properties": {
          "type_code": {
            "type": "string",
            "minLength": 1,
            "title": "Type Code"
          },
          "label": {
            "type": "string",
            "minLength": 1,
            "title": "Label"
          },
          "value": {
            "type": "string",
            "minLength": 1,
            "title": "Value"
          },
          "order_no": {
            "type": "integer",
            "title": "Order No",
            "default": 0
          }
        },
        "type": "object",
        "required": [
          "type_code",
          "label",
          "value"
        ],
        "title": "DictItemCreate"
      },
      "DictItemOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "type_code": {
            "type": "string",
            "title": "Type Code"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "value": {
            "type": "string",
            "title": "Value"
          },
          "order_no": {
            "type": "integer",
            "title": "Order No",
            "default": 0
          },
          "status": {
            "type": "string",
            "title": "Status",
            "default": "active"
          }
        },
        "type": "object",
        "required": [
          "id",
          "type_code",
          "label",
          "value"
        ],
        "title": "DictItemOut"
      },
      "DictItemUpdate": {
        "properties": {
          "label": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Label"
          },
          "value": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Value"
          },
          "order_no": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Order No"
          },
          "status": {
            "anyOf": [
              {
                "type": "string",
                "enum": [
                  "active",
                  "disabled"
                ]
              },
              {
                "type": "null"
              }
            ],
            "title": "Status"
          }
        },
        "type": "object",
        "title": "DictItemUpdate"
      },
      "DictOut": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "code": {
            "type": "string",
            "title": "Code"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "remark": {
            "type": "string",
            "title": "Remark",
            "default": ""
          }
        },
        "type": "object",
        "required": [
          "id",
          "code",
          "name"
        ],
        "title": "DictOut"
      },
      "DictUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "remark": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Remark"
          }
        },
        "type": "object",
        "title": "DictUpdate"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}